requirements.txt
torch>=2.0.0
transformers>=4.35.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
numpy>=1.21.0
